🔨 [functional] Move `installroot` tests to separate module
