+++ tests/functional/test_installroot.py
@@ -1,0 +1,100 @@
+"""Functional tests for ``installroot``."""
+import nox.sessions
+from tests.functional.conftest import list_packages
+from tests.functional.conftest import Project
+from tests.functional.conftest import run_nox_with_noxfile
+
+import nox_poetry
+
+
+def test_installroot_no_install(project: Project) -> None:
+    """It skips installation when --no-install is passed."""
+
+    @nox_poetry.session
+    def test(session: nox_poetry.Session) -> None:
+        """Install the local package."""
+        session.poetry.installroot()
+
+    run_nox_with_noxfile(project, [test], [nox_poetry])
+    run_nox_with_noxfile(project, [test], [nox_poetry], "-R")
+
+    expected = [project.package, *project.dependencies]
+    packages = list_packages(project, test)
+
+    assert set(expected) == set(packages)
+
+
+def test_installroot_wheel(project: Project) -> None:
+    """It builds and installs a wheel from the local package."""
+
+    @nox.session
+    def test(session: nox.sessions.Session) -> None:
+        """Install the local package."""
+        nox_poetry.installroot(session, distribution_format=nox_poetry.WHEEL)
+
+    run_nox_with_noxfile(project, [test], [nox, nox.sessions, nox_poetry])
+
+    expected = [project.package, *project.dependencies]
+    packages = list_packages(project, test)
+
+    assert set(expected) == set(packages)
+
+
+def test_installroot_wheel_with_extras(project: Project) -> None:
+    """It installs the extra."""
+
+    @nox.session
+    def test(session: nox.sessions.Session) -> None:
+        """Install the local package."""
+        nox_poetry.installroot(
+            session, distribution_format=nox_poetry.WHEEL, extras=["pygments"]
+        )
+
+    run_nox_with_noxfile(project, [test], [nox, nox.sessions, nox_poetry])
+
+    expected = [
+        project.package,
+        *project.dependencies,
+        project.get_dependency("pygments"),
+    ]
+    packages = list_packages(project, test)
+
+    assert set(expected) == set(packages)
+
+
+def test_installroot_sdist(project: Project) -> None:
+    """It builds and installs an sdist from the local package."""
+
+    @nox.session
+    def test(session: nox.sessions.Session) -> None:
+        """Install the local package."""
+        nox_poetry.installroot(session, distribution_format=nox_poetry.SDIST)
+
+    run_nox_with_noxfile(project, [test], [nox, nox.sessions, nox_poetry])
+
+    expected = [project.package, *project.dependencies]
+    packages = list_packages(project, test)
+
+    assert set(expected) == set(packages)
+
+
+def test_installroot_sdist_with_extras(project: Project) -> None:
+    """It installs the extra."""
+
+    @nox.session
+    def test(session: nox.sessions.Session) -> None:
+        """Install the local package."""
+        nox_poetry.installroot(
+            session, distribution_format=nox_poetry.SDIST, extras=["pygments"]
+        )
+
+    run_nox_with_noxfile(project, [test], [nox, nox.sessions, nox_poetry])
+
+    expected = [
+        project.package,
+        *project.dependencies,
+        project.get_dependency("pygments"),
+    ]
+    packages = list_packages(project, test)
+
+    assert set(expected) == set(packages)
